requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
black==23.12.0
certifi==2023.11.17
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
fastapi==0.105.0
fonttools==4.46.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
isort==5.13.1
kiwisolver==1.4.5
matplotlib==3.8.2
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pytest==7.4.3
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
yarl==1.9.4
